Refactor services to public and move to Services namespace
Refactored MyEnvironment, NativeNavigationManager, SessionStorage, and NativeSettings to the Promise.Native.Services namespace and changed their access modifiers to public for broader accessibility. Updated MauiProgram.cs to use the new namespace. SessionStorage now uses a primary constructor and applies .ConfigureAwait(true) to async JS interop calls. Simplified .csproj NoWarn settings to only suppress CA1515.
