+++ Promise.Native/Services/SessionStorage.cs
@@ -1,34 +1,28 @@
 ﻿using Microsoft.JSInterop;
 
-namespace Promise.Native;
+namespace Promise.Native.Services;
 
-#pragma warning disable CA2007 // IJSRuntime requires synchronization context; ConfigureAwait(false) would break it
-internal sealed class SessionStorage : ISessionStorage
+public sealed class SessionStorage(IJSRuntime jsRuntime) : ISessionStorage
 {
-    private readonly IJSRuntime _jsRuntime;
-
-    public SessionStorage(IJSRuntime jsRuntime)
-    {
-        _jsRuntime = jsRuntime;
-    }
+    private readonly IJSRuntime _jsRuntime = jsRuntime;
 
     public async Task<T> GetAsync<T>(string key)
     {
-        return await _jsRuntime.InvokeAsync<T>("localStorage.getItem", key);
+        return await _jsRuntime.InvokeAsync<T>("localStorage.getItem", key).ConfigureAwait(true);
     }
 
     public async Task SetAsync<T>(string key, T value)
     {
-        await _jsRuntime.InvokeVoidAsync("localStorage.setItem", key, value);
+        await _jsRuntime.InvokeVoidAsync("localStorage.setItem", key, value).ConfigureAwait(true);
     }
 
     public async Task RemoveAsync(string key)
     {
-        await _jsRuntime.InvokeVoidAsync("localStorage.removeItem", key);
+        await _jsRuntime.InvokeVoidAsync("localStorage.removeItem", key).ConfigureAwait(true);
     }
 
     public async Task ClearAsync()
     {
-        await _jsRuntime.InvokeVoidAsync("localStorage.clear");
+        await _jsRuntime.InvokeVoidAsync("localStorage.clear").ConfigureAwait(true);
     }
 }
